server/requirements.txt
py-cord
python-dotenv
websockets
uuid